requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tqdm==4.66.5
python-docx==1.1.2
PyMuPDF==1.24.9
numpy==2.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
python-dotenv==1.0.1
pydantic==2.8.2